+++ 03_Printf/ft_tmp2buff.c
@@ -6,7 +6,7 @@
 /*   By: tderwedu <tderwedu@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/26 17:38:21 by tderwedu          #+#    #+#             */
-/*   Updated: 2021/03/02 15:09:42 by tderwedu         ###   ########.fr       */
+/*   Updated: 2021/03/04 11:38:17 by tderwedu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_tmp2buff(t_vec *buff, t_vec *tmp)
 {
-	size_t len;
+	int len;
 
 	len = tmp->len;
 	while (len > (buff->max - buff->ptr))
